% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{add_forecast_results}
\alias{add_forecast_results}
\title{After making predictions the week prior, add the actual toxicity classification to assess}
\usage{
add_forecast_results()
}
\description{
After making predictions the week prior, add the actual toxicity classification to assess
}
