% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{append_forecast}
\alias{append_forecast}
\title{Adds new forecast data to the database}
\usage{
append_forecast(predictions)
}
\arguments{
\item{predictions}{tibble of new shellfish toxicity classification predictions}
}
\description{
Adds new forecast data to the database
}
