% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{empty_forecast}
\alias{empty_forecast}
\title{Defines an empty forecast table}
\usage{
empty_forecast(which = c("single", "ensemble")[1])
}
\arguments{
\item{which}{character selecting single or ensemble forecast template}
}
\value{
tibble with forecast columns and no rows
}
\description{
Defines an empty forecast table
}
