% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closures.R
\name{find_closures}
\alias{find_closures}
\title{Finds the date of the date of the peak toxicity measurement and the date it was predicted by the forecast}
\usage{
find_closures(psp, predictions)
}
\arguments{
\item{psp}{tibble of psp observations}

\item{predictions}{tibble of psp forecast predictions}
}
\value{
a table with one row per site
}
\description{
Finds the date of the date of the peak toxicity measurement and the date it was predicted by the forecast
}
