% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_metrics.R
\name{find_station_metrics}
\alias{find_station_metrics}
\title{Finds skill metrics for each station included in the experimental forecast}
\usage{
find_station_metrics(results = read_all_results())
}
\arguments{
\item{results}{tibble of results with columns class for truth and predicted class for estimate}
}
\value{
tibble of metrics with one row per station
}
\description{
Finds skill metrics for each station included in the experimental forecast
}
