% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_metrics.R
\name{forecast_metrics}
\alias{forecast_metrics}
\title{Finds metrics from a table of forecast predictions}
\usage{
forecast_metrics(
  fc,
  predicted_col = "predicted_class",
  measured_col = "actual_class"
)
}
\arguments{
\item{fc}{tibble of predictions}

\item{predicted_col}{character string of predicted classification column name}

\item{measured_col}{character string of measured classification column}
}
\description{
Finds metrics from a table of forecast predictions
}
