% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{format_probs}
\alias{format_probs}
\title{Takes a column of probabilities and formats them for stakeholder delivery (rounds to integer and adds percent symbol)}
\usage{
format_probs(probs)
}
\arguments{
\item{probs}{vector of probabilities}
}
\value{
formatted vector of probabilities
}
\description{
Takes a column of probabilities and formats them for stakeholder delivery (rounds to integer and adds percent symbol)
}
