% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_webpage_table.R
\name{format_webpage_table}
\alias{format_webpage_table}
\title{Formats a psp forecast table for the DMR closure map webpage}
\usage{
format_webpage_table(f)
}
\arguments{
\item{f}{tibble of psp forecast predictions}
}
\value{
formatted tibble
}
\description{
Formats a psp forecast table for the DMR closure map webpage
}
