% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_year.R
\name{get_recent_year}
\alias{get_recent_year}
\title{Checks the most recent year in the forecast directory in github.com/pspforecast/inst/forecastdb}
\usage{
get_recent_year(earliest = 2021)
}
\arguments{
\item{earliest}{numeric defining the earliest expected year in the directory}
}
\value{
a character year of the most recent prediction file
}
\description{
Checks the most recent year in the forecast directory in github.com/pspforecast/inst/forecastdb
}
