% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{read_all_predictions}
\alias{read_all_predictions}
\title{Reads all predictions from all years into one tibble}
\usage{
read_all_predictions(years = 2021:2023)
}
\arguments{
\item{years}{numeric vector of years}
}
\value{
tibble with all predictions made
}
\description{
Reads all predictions from all years into one tibble
}
