% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_all_results}
\alias{read_all_results}
\title{Reads all results in seasonal_results directory}
\usage{
read_all_results(years = 2021:2023)
}
\arguments{
\item{years}{vector of integer years}
}
\value{
tibble of results
}
\description{
Reads all results in seasonal_results directory
}
