% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_results}
\alias{read_results}
\title{Reads experimental predictions matched with their outcome for one year}
\usage{
read_results(year = c(2021, 2022, 2023, 2024, 2025)[1])
}
\arguments{
\item{year}{integer year to read}
}
\value{
tibble of results
}
\description{
Reads experimental predictions matched with their outcome for one year
}
