% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_classification.R
\name{recode_classification}
\alias{recode_classification}
\title{Set classification values}
\usage{
recode_classification(v, lut = c(0, 10, 30, 80), na_value = 0)
}
\arguments{
\item{v}{vector of numbers (x$total_toxicity)}

\item{lut}{ordered vector of toxicity levels (tox_levels)}

\item{na_value}{value to replace missing values in v}
}
\value{
ix vector closure codes
}
\description{
Set classification values
}
