% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{write_forecast}
\alias{write_forecast}
\title{Adds a new forecast file of predictions/data to the database}
\usage{
write_forecast(predictions)
}
\arguments{
\item{predictions}{tibble of new shellfish toxicity classification predictions}
}
\description{
Adds a new forecast file of predictions/data to the database
}
