% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_webpage_table.R
\name{write_forecast_tables}
\alias{write_forecast_tables}
\title{Writes two table containing the current forecast. One is formatted for the Interactive Shellfish Closure 
Map and the other has all columns and no formatting.}
\usage{
write_forecast_tables(user_config)
}
\arguments{
\item{user_config}{list of pspforecast user configuration containing paths to users local copies of both tables}
}
\description{
Writes two table containing the current forecast. One is formatted for the Interactive Shellfish Closure 
Map and the other has all columns and no formatting.
}
