% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_forecast_results.R
\name{add_forecast_results}
\alias{add_forecast_results}
\title{After making predictions the week prior, add the actual toxicity classification to assess}
\usage{
add_forecast_results(
  predictions,
  toxin_measurements,
  tox_levels = c(0, 10, 30, 80)
)
}
\arguments{
\item{predictions}{table of predictions (from pspforecast::read_forecast())}

\item{toxin_measurements}{table of psp data}

\item{tox_levels}{vector of toxicity classification bins}
}
\value{
table of predictions with the closest matching toxicity measurement
}
\description{
After making predictions the week prior, add the actual toxicity classification to assess
}
