% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{read_forecast}
\alias{read_forecast}
\title{Reads forecast database}
\usage{
read_forecast()
}
\value{
tibble of forecasted shellfish toxicity classifications along with their metadata
}
\description{
Reads forecast database
}
