% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{read_forecast}
\alias{read_forecast}
\title{Reads forecast database}
\usage{
read_forecast(
  format = FALSE,
  new_only = FALSE,
  shiny = FALSE,
  id = FALSE,
  year = get_recent_year()
)
}
\arguments{
\item{format}{logical, if true, the forecast report will be formatted for stakeholders with rounded probabilities and 0 probabilities being changed to <1}

\item{new_only}{logical, if true, then only the newest observations from each station will be served}

\item{shiny}{logical, if true, forecast will be read from github csv rather than local; method used for deploying package in shiny app server}

\item{id}{logical if true the tibble of predictions returned will have an f_id column that is the location and date pasted together}

\item{year}{character string selecting which year's predictions to read - default is to find the most recent year in the forecastdb directory}
}
\value{
tibble of predicted shellfish toxicity classifications along with their metadata
}
\description{
Reads forecast database
}
