% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_io.R
\name{read_forecast}
\alias{read_forecast}
\title{Reads forecast database}
\usage{
read_forecast(new_only = FALSE)
}
\arguments{
\item{new_only}{logical, if true then only the newest observations from each station will be served}
}
\value{
tibble of predicted shellfish toxicity classifications along with their metadata
}
\description{
Reads forecast database
}
